refactor: replace select with universal dropdown
