+++ src/shared/ui/dropdown/ui.module.scss
@@ -1,0 +1,23 @@
+.root {
+  position: relative;
+}
+
+.root[data-disabled='true'] {
+  opacity: 0.6;
+}
+
+.dropdown {
+  position: absolute;
+  left: 0;
+  top: calc(100% + 8px);
+  width: 100%;
+  background: #fff;
+  border-radius: 10px;
+  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.12);
+  z-index: 50;
+  overflow: hidden;
+}
+
+.dropdownInner {
+  padding: 4px;
+}
