--- src/shared/ui/select/index.ts
@@ -1,2 +1,0 @@
-export { Select } from './ui'
-export type { SelectItem } from './ui'
