--- src/shared/ui/select/ui.module.scss
@@ -1,64 +1,0 @@
-@use '@styles/variables' as *;
-
-.root {
-  position: relative;
-}
-
-.root[data-disabled='true'] {
-  opacity: 0.6;
-}
-
-.dropdown {
-  position: absolute;
-  left: 0;
-  top: calc(100% + 8px);
-  width: 100%;
-  background: #fff;
-  border-radius: 10px;
-  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.12);
-  z-index: 50;
-  overflow: hidden;
-}
-
-.dropdownInner {
-  padding: 4px;
-}
-
-.item {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 7px 10px;
-  border-radius: 8px;
-  background: transparent;
-  border: none;
-  cursor: pointer;
-  text-align: left;
-  color: $text-heading;
-  transition:
-    background-color 0.2s ease,
-    color 0.2s ease;
-}
-
-.item:hover {
-  background: rgba(0, 0, 0, 0.04);
-}
-
-.itemSelected {
-  color: $color-primary;
-}
-
-.itemIcon {
-  width: 16px;
-  height: 16px;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  flex: 0 0 16px;
-}
-
-.itemText {
-  flex: 1 1 auto;
-  min-width: 0;
-}
